requirements.txt
async-timeout==4.0.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==2.1.3
Flask-Cors==4.0.0
Flask-SeaSurf==1.1.1
flask-talisman==1.1.0
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-datastore==2.19.0
google-cloud-ndb==2.3.1
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
gunicorn==21.2.0
idna==3.6
importlib-metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
pymemcache==4.0.0
pytz==2024.1
redis==5.0.3
requests==2.31.0
rsa==4.9
urllib3==2.2.1
Werkzeug==2.2.2
zipp==3.18.1
